+++ examples/kafkajs/admin/describe-groups.js
@@ -1,5 +1,5 @@
 // require('kafkajs') is replaced with require('@confluentinc/kafka-javascript').KafkaJS.
-const { Kafka, ConsumerGroupStates } = require('@confluentinc/kafka-javascript').KafkaJS;
+const { Kafka } = require('@confluentinc/kafka-javascript').KafkaJS;
 const { parseArgs } = require('node:util');
 
 function printNode(node, prefix = '') {
@@ -72,6 +72,7 @@ async function adminStart() {
       console.log(`\tProtocol type: ${group.protocolType}`);
       console.log(`\tPartition assignor: ${group.partitionAssignor}`);
       console.log(`\tState: ${group.state}`);
+      console.log(`\tType: ${group.type}`);
       console.log(`\tCoordinator: ${group.coordinator ? group.coordinator.id : group.coordinator}`);
       printNode(group.coordinator, '\t');
       console.log(`\tAuthorized operations: ${group.authorizedOperations}`);